+++ asciidoc_dita_toolkit/asciidoc_dita/cli_utils.py
@@ -1,0 +1,34 @@
+"""
+cli_utils.py - CLI argument parsing utilities.
+
+This module provides reusable CLI argument parsing functionality for AsciiDoc processing tools.
+"""
+
+import argparse
+
+
+def common_arg_parser(parser):
+    """
+    Add standard options to the supplied parser:
+    -d / --directory: Root directory to search (default: current directory)
+    -r / --recursive: Search subdirectories recursively
+    -f / --file: Scan only the specified .adoc file
+
+    Args:
+        parser: ArgumentParser instance to add arguments to
+    """
+    sources = parser.add_mutually_exclusive_group()
+    sources.add_argument(
+        "-d",
+        "--directory",
+        type=str,
+        default=".",
+        help="Root directory to search (default: current directory)",
+    )
+    sources.add_argument("-f", "--file", type=str, help="Scan only the specified .adoc file")
+    parser.add_argument(
+        "-r",
+        "--recursive",
+        action="store_true",
+        help="Search subdirectories recursively",
+    )
